bump golang to 1.23.3 and golangci-lint to 1.60.3
Bump golang to 1.23.3 and golangci-lint to 1.60.3 which supports Go
1.23.

Get rid of deprecated golangci-lint configs, and add timeout to the
action, otherwise it'll be timed out in 1 minute. 15 minutes is
consistent with "make lint" timeout.

Linter fixes:
- Fix new linter errors by removing the return value which is always
nil (unparam linter), and thus meaningless.
- "exportloopref" is not needed from Go 1.22 onwards, replaced by
"copyloopvar" in golangci-lint 1.60.2.
- Ambiguous string warnings in Wrapf fixed.

+++ test/e2e/remediation.go
@@ -136,7 +136,7 @@ func remediation(ctx context.Context, inputGetter func() RemediationInput) {
 
 	By("Testing unhealthy and inspection annotations")
 	By("Scaling down KCP to 1 replica")
-	newReplicaCount := 1
+	newReplicaCount := int32(1)
 	ScaleKubeadmControlPlane(ctx, bootstrapClient, client.ObjectKey{Namespace: "metal3", Name: "test1"}, newReplicaCount)
 	WaitForNumBmhInState(ctx, bmov1alpha1.StateAvailable, WaitForNumInput{
 		Client:    bootstrapClient,
